test(iast): fix the fallout from moving appsec servers to free ports DD_7MBVTY DD_PRPLYS DD_L13XP8 DD_TFY7YU DD_D5HQHX DD_CQB771 DD_WY4GDW DD_KUBWIR DD_JU041J DD_FQAL2Q DD_AY22HP DD_3J27UL DD_8A0ANP DD_LK4TUP DD_SNBL4W DD_D07T24 DD_6TE3BU DD_NUI8G4
Two breakages from the previous commit:

The shared integrations conftest imported get_free_port from appsec_utils,
which imports psutil. appsec_integrations_flask runs the only two files under
that conftest that never imported appsec_utils, and its venv has no psutil, so
every shard died at collection. get_free_port and port_is_available move to
tests/appsec/ports.py, which imports nothing but socket.

The MCP tests build their SSE client URL from a port passed separately from the
server's, and those four call sites still had the old literals, so the client
dialled 8052/8053 while the server listened elsewhere.

ports.py is declared alongside appsec_utils.py in suitespec so a change to it
still triggers the suites that depend on it.
