+++ tests/appsec/suitespec.yml
@@ -76,6 +76,7 @@ suites:
       - tests/appsec/iast_packages/*
       - tests/appsec/app.py
       - tests/appsec/appsec_utils.py
+      - tests/appsec/ports.py
     timeout: 50m
   iast_tdd_propagation:
     venvs_per_job: 1
@@ -88,6 +89,7 @@ suites:
       - tests/appsec/iast_tdd_propagation/*
       - tests/appsec/app.py
       - tests/appsec/appsec_utils.py
+      - tests/appsec/ports.py
     retry: 2
     snapshot: true
   appsec_integrations_pygoat:
@@ -154,6 +156,7 @@ suites:
       - tests/appsec/integrations/flask_tests/test_appsec_flask_telemetry.py
       - tests/appsec/app.py
       - tests/appsec/appsec_utils.py
+      - tests/appsec/ports.py
     retry: 2
     # test_appsec_flask_telemetry.py asserts on payloads received by the test agent.
     snapshot: true
@@ -170,6 +173,7 @@ suites:
       - tests/appsec/integrations/flask_tests/*
       - tests/appsec/app.py
       - tests/appsec/appsec_utils.py
+      - tests/appsec/ports.py
     retry: 2
     services:
       - testagent
@@ -185,6 +189,7 @@ suites:
       - '@remoteconfig'
       - tests/appsec/integrations/django_tests/*
       - tests/appsec/appsec_utils.py
+      - tests/appsec/ports.py
     retry: 2
     services:
       - testagent
@@ -200,6 +205,7 @@ suites:
       - '@remoteconfig'
       - tests/appsec/integrations/fastapi_tests/*
       - tests/appsec/appsec_utils.py
+      - tests/appsec/ports.py
     retry: 2
     services:
       - testagent
